+++ megamek/src/megamek/client/bot/caspar/axis/MovingTowardWaypointCalculator.java
@@ -37,9 +37,9 @@ import megamek.common.Coords;
  */
 public class MovingTowardWaypointCalculator extends BaseAxisCalculator {
     @Override
-    public double[] calculateAxis(Pathing pathing, GameState gameState) {
+    public float[] calculateAxis(Pathing pathing, GameState gameState) {
         // This calculates if the unit is moving toward the waypoint
-        double[] movingTowardWaypoint = axis();
+        float[] movingTowardWaypoint = axis();
         if (pathing.hasWaypoint()) {
             Coords waypoint = pathing.getWaypoint();
             int distFromStart = pathing.getStartCoords().distance(waypoint);
