+++ megamek/src/megamek/client/bot/caspar/axis/ScoutingCalculator.java
@@ -36,9 +36,9 @@ import megamek.client.bot.common.Pathing;
  */
 public class ScoutingCalculator extends BaseAxisCalculator {
     @Override
-    public double[] calculateAxis(Pathing pathing, GameState gameState) {
+    public float[] calculateAxis(Pathing pathing, GameState gameState) {
         // This calculates the scouting
-        double[] scouting = axis();
+        float[] scouting = axis();
 
         return scouting;
     }
