+++ megamek/src/megamek/client/bot/caspar/axis/UnitTmmCalculator.java
@@ -36,9 +36,9 @@ import megamek.client.bot.common.Pathing;
  */
 public class UnitTmmCalculator extends BaseAxisCalculator {
     @Override
-    public double[] calculateAxis(Pathing pathing, GameState gameState) {
+    public float[] calculateAxis(Pathing pathing, GameState gameState) {
         // This calculates the unit TMM
-        double[] unitTmm = axis();
+        float[] unitTmm = axis();
 
         return unitTmm;
     }
